new: Variadic arguments Functions can now accept variable number of arguments from the user using the syntax fn doSomething(x, y..) {     // x is necessary argument     // y contains the varargs in a tuple } Next does not limit maximum number of necessary arguments, but it does limit maximum number of variadic arguments, by default to 32, since Next supports overloading, and a variadic function has to be registered for all possible invocations of it. Except for the signature, all of the symbols in the table points to the same Function*. Variadic functions can even be referred using the '@' syntax like normal methods. Since Next already had the variadic backend somewhat in place, this commit mostly enables frontend changes, and additionally, now the fiber takes care of contracting the additional arguments in a tuple if the callee is variadic. If the callee is a builtin, nothing is changed, since we already pass the argcount to builtins. This commit also fixes a few things: 1. Engine now prints the stacktrace _after_ printing the exception. 2. 'load_module' did not check for the object at slot 0 to be a class,    and by default assumed that it is. This is now fixed. 3. A new sentinel 'end' opcode is introduced, which marks the end of the    bytecode instructions. 4. The IDXERR macro now requires the second argument to be a string at    compile time, instead of making it a string by itself. 5. Array and Tuple now output a clear exception message when empty. 6. IndexError also prints the associated message provided by the    thrower. 7. Class.add_builtin_fn now supports adding a variadic builtin, by    passing the last argument as 'true'. 8. core.print and core.fmt are now implemented as true vararg functions    now. Since vararg functions have a limit on their max arg count, this    also somewhat directly limits the argcount of them. But the default    is set to be 32 for now, and so it shouldn't be much of a problem. 9. A new builtin core.println is now added, which does as advertised. 10. Array.insert and String.append are now also variadic. 11. Performance remains largely the same, except fib is showing 6%     improvements somehow.

+++ objects/bytecode.h
@@ -25,7 +25,7 @@ struct Bytecode {
 	BytecodeCompilationContext *ctx; // debug info
 	size_t                      size;
 	size_t                      capacity;
-	size_t                      numSlots;
+	int                         numSlots;
 	int                         stackSize;
 	int                         stackMaxSize;
